feat: added symbolicImage field & use only one asset resolver chain
