+++ src/Resolver/Asset/ResourceAssetResolverChain.php
@@ -4,51 +4,27 @@ declare(strict_types=1);
 
 namespace Atoolo\GraphQL\Search\Resolver\Asset;
 
-use Atoolo\GraphQL\Search\Resolver\Resolver;
 use Atoolo\GraphQL\Search\Types\Asset;
+use Atoolo\GraphQL\Search\Types\SymbolicImage;
 use Atoolo\Resource\Resource;
 use Overblog\GraphQLBundle\Definition\ArgumentInterface;
 
-class ResourceAssetResolverChain implements Resolver
+class ResourceAssetResolverChain implements ResourceAssetResolver, ResourceSymbolicImageResolver
 {
     /**
-     * @param array<ResourceAssetResolver> $regularAssetResolvers
-     * @param array<ResourceAssetResolver> $symbolicAssetResolvers
+     * @param array<ResourceAssetResolver> $resolvers
      */
     public function __construct(
-        private readonly iterable $regularAssetResolvers,
-        private readonly iterable $symbolicAssetResolvers
+        private readonly iterable $resolvers
     ) {
     }
 
-    /**
-     * Iterates over all regular asset resolvers (tagged with
-     * `atoolo_graphql_search.resolver.asset.regular`) and all
-     * symbolic asset resolvers (tagged with
-     * `atoolo_graphql_search.resolver.asset.symbolic`) and returns
-     * the first resolved asset that is not null. All  regular asset
-     * resolvers will always have priority over all symbolic asset resolvers.
-     * Passing the argument `forceSymbolic` will cause this method to
-     * only apply symbolic asset resolvers, ignoring the regular assert
-     * resolvers.
-     */
     public function getAsset(
         Resource $resource,
         ArgumentInterface $args
     ): ?Asset {
-        if ($args['forceSymbolic'] ?? false) {
-            return $this->getAssetSymbolic($resource, $args);
-        }
-        return $this->getAssetRegular($resource, $args)
-            ?? $this->getAssetSymbolic($resource, $args);
-    }
-
-    public function getAssetRegular(
-        Resource $resource,
-        ArgumentInterface $args
-    ): ?Asset {
-        foreach ($this->regularAssetResolvers as $regularAssetResolver) {
-            $asset = $regularAssetResolver->getAsset(
+        foreach ($this->resolvers as $resolver) {
+            $asset = $resolver->getAsset(
                 $resource,
                 $args
             );
@@ -59,17 +35,20 @@ class ResourceAssetResolverChain implements Resolver
         return null;
     }
 
-    public function getAssetSymbolic(
+    public function getSymbolicImage(
         Resource $resource,
         ArgumentInterface $args
-    ): ?Asset {
-        foreach ($this->symbolicAssetResolvers as $symbolicAssetResolvers) {
-            $asset = $symbolicAssetResolvers->getAsset(
+    ): ?SymbolicImage {
+        foreach ($this->resolvers as $resolver) {
+            if (!$resolver instanceof ResourceSymbolicImageResolver) {
+                continue;
+            }
+            $symbolicImage = $resolver->getSymbolicImage(
                 $resource,
                 $args
             );
-            if ($asset !== null) {
-                return $asset;
+            if ($symbolicImage !== null) {
+                return $symbolicImage;
             }
         }
         return null;
