+++ src/Resolver/NewsTeaserResolver.php
@@ -4,16 +4,19 @@ declare(strict_types=1);
 
 namespace Atoolo\GraphQL\Search\Resolver;
 
-use Atoolo\GraphQL\Search\Resolver\Asset\ResourceAssetResolverChain;
+use Atoolo\GraphQL\Search\Resolver\Asset\ResourceAssetResolver;
+use Atoolo\GraphQL\Search\Resolver\Asset\ResourceSymbolicImageResolver;
 use Atoolo\GraphQL\Search\Types\Asset;
 use Atoolo\GraphQL\Search\Types\NewsTeaser;
+use Atoolo\GraphQL\Search\Types\SymbolicImage;
 use DateTime;
 use Overblog\GraphQLBundle\Definition\ArgumentInterface;
 
 class NewsTeaserResolver implements Resolver
 {
     public function __construct(
-        private readonly ResourceAssetResolverChain $assetResolver,
+        private readonly ResourceAssetResolver $assetResolver,
+        private readonly ResourceSymbolicImageResolver $symbolicImageResolver,
         private readonly ResourceDateResolver $dateResolver
     ) {
     }
@@ -29,5 +32,12 @@ class NewsTeaserResolver implements Resolver
         ArgumentInterface $args
     ): ?Asset {
         return $this->assetResolver->getAsset($teaser->resource, $args);
+    }
+
+    public function getSymbolicImage(
+        NewsTeaser $teaser,
+        ArgumentInterface $args
+    ): ?SymbolicImage {
+        return $this->symbolicImageResolver->getSymbolicImage($teaser->resource, $args);
     }
 }
